requirements.txt
torch
torchvision
numpy
scikit-learn
Pillow
gradio
matplotlib
tqdm